Moved apiCalming into common folder.
