+++ src/js/pages/Ballot/Candidate.jsx
@@ -28,7 +28,7 @@ import VoterStore from '../../stores/VoterStore';
 import { isAndroidSizeFold } from '../../common/utils/cordovaUtils';
 import { isWebApp } from '../../common/utils/isCordovaOrWebApp';
 import { renderLog } from '../../common/utils/logging';
-import apiCalming from '../../utils/apiCalming';
+import apiCalming from '../../common/utils/apiCalming';
 import { PageContentContainer } from '../../utils/pageLayoutStyles';
 import { capitalizeString, convertToInteger } from '../../utils/textFormat';
