+++ src/js/pages/ReadyLight.jsx
@@ -21,7 +21,7 @@ import { isWebApp } from '../common/utils/isCordovaOrWebApp';
 import historyPush from '../common/utils/historyPush';
 import lazyPreloadPages from '../utils/lazyPreloadPages';
 import { renderLog } from '../common/utils/logging';
-import apiCalming from '../utils/apiCalming';
+import apiCalming from '../common/utils/apiCalming';
 import { PageContentContainer } from '../utils/pageLayoutStyles';
 
 const ReadMore = React.lazy(() => import(/* webpackChunkName: 'ReadMore' */ '../common/components/Widgets/ReadMore'));
